requirements.txt
numpy
scikit-learn
termcolor
regex
tqdm
prettytable
scipy
nltk
elasticsearch
pexpect==4.2.1
pandas
mecab-python3